backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1
PyMuPDF==1.25.3
python-docx==1.1.2
spacy==3.8.4
keybert==0.8.5
rapidfuzz==3.11.0
sentence-transformers==3.3.1
scikit-learn==1.6.1
google-generativeai==0.8.4
sqlalchemy==2.0.36
aiosqlite==0.20.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
